Fix CRITICAL bug: Lesson.total_cost() now returns Decimal not float
Bug found by unit tests (test_lesson_total_cost_uses_decimal_precision):
- total_cost() was converting Decimal calculation to float
- This could cause floating point precision errors with money
- Now returns Decimal directly for exact precision

Example: 1.5 hours × $80/hr = $120.00 (exact), not 120.000000001

This is a HIGH PRIORITY fix - affects all invoice calculations.

+++ billing/models.py
@@ -122,7 +122,8 @@ class Lesson(models.Model):
         # Ensure both values are Decimal for proper calculation
         rate = Decimal(str(self.rate)) if not isinstance(self.rate, Decimal) else self.rate
         duration = Decimal(str(self.duration)) if not isinstance(self.duration, Decimal) else self.duration
-        return float(rate * duration)
+        # CRITICAL: Return Decimal for money precision, not float
+        return rate * duration
     
     def save(self, *args, **kwargs):
         # Set rate from teacher's hourly rate if not provided and teacher has a custom rate
